requirement.txt
atomicwrites==1.4.0
attrs==21.4.0
biopython==1.79
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
codecov==2.1.12
colorama==0.4.4
coverage==6.3.2
Flask==2.0.3
Flask-SQLAlchemy==2.5.1
glob2==0.7
greenlet==1.1.2
idna==3.3
iniconfig==1.1.1
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
more-itertools==8.12.0
mysql==0.0.3
mysql-connector==2.2.9
mysql-connector-python==8.0.28
mysqlclient==2.1.0
numpy==1.22.2
oyaml==1.0
packaging==21.3
pandas==1.4.1
pluggy==1.0.0
protobuf==3.19.4
py==1.11.0
pyparsing==3.0.7
pytest==7.0.1
pytest-cov==3.0.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
SQLAlchemy==1.4.32
tomli==2.0.1
urllib3==1.26.8
Werkzeug==2.0.3
xmljson==0.2.1
